Move all interfaces into one folder
