+++ src/interfaces/ICloseCommandArgs.ts
