+++ src/interfaces/ICommand.ts
